requirements.txt
Flask
Flask-PyMongo
Flask-Login
Flask-SocketIO
Flask-WTF
Werkzeug
certifi
python-dotenv
